requirements.txt
certifi
charset-normalizer
idna
PyYAML
requests
six
transliterate
typing_extensions
urllib3
numpy
torch
